+++ app/server/src/middleware/index.ts
@@ -1,8 +1,7 @@
 import { type RouterTypes } from 'bun';
 import { type WSMsg, read, write } from '@tbsp/web/ws/protocol.ts';
-import { HTTPRegister, RouteRegister } from '@tbsp/web';
+import { HTTPRegister } from '@tbsp/web';
 import WebSocketRegister, { type TbspWebSocketHandler } from '@tbsp/web/ws';
-import fs from 'node:fs';
 
 type HTTPMethod = 'GET' | 'POST' | 'PUT' | 'DELETE';
 type RoutePath = string;
@@ -28,54 +27,6 @@ function respondOrUpgrade(handler: RouteHandler): RouteHandler {
     if (isUpgrade === 'websocket') return upgrade(req, server);
     return handler(req, server);
   };
-}
-
-function debugPath<T>(path: string, fallback: T) {
-  try {
-    fs.accessSync(path);
-  } catch {
-    console.error(`Failed to access file at ${path} (resolved from ${process.cwd()}).`);
-    return fallback;
-  }
-}
-
-/**
- * Serve a response from a file at `path`.
- * This uses the environment variable `DEV` (development mode)
- * to define its behaviour. (It's a wrapper around `Bun.file`)
- * - in development: read the file and send the content
- * - in production: load the file into memory immediately and send content on request without fs IO
- */
-export function File(path: string, preload = false) {
-  let r = debugPath(path, () => new Response('Resource not found.', { status: 404 }));
-  if (r) return r;
-
-  if (preload) {
-    return async () => new Response(await Bun.file(path).bytes());
-  } else {
-    return () => new Response(Bun.file(path));
-  }
-}
-
-export function PublicDirectory(basePath: string, preload = false, prefix = '') {
-  basePath = basePath.endsWith('/') ? basePath.slice(0, basePath.length - 1) : basePath;
-  let app = new TBSPApp();
-
-  let r = debugPath(basePath, app);
-  if (r) return r;
-
-  for (let item of fs.readdirSync(basePath)) {
-    let itemPath = `${basePath}/${item}`;
-    if (debugPath(itemPath, true)) continue;
-
-    if (fs.statSync(itemPath).isDirectory()) {
-      app.use(PublicDirectory(`${basePath}/${item}`, preload, `${item}/`));
-    } else {
-      // console.log(`Registered file ${itemPath} at /${prefix}${item}`);
-      app.get(`/${prefix}${item}`, File(itemPath, preload));
-    }
-  }
-  return app;
 }
 
 function displayObject(obj: object, tabWidth = 4): string {
@@ -235,7 +186,7 @@ export default class TBSPApp {
   }
 
   getCompiledWebSocket(): Bun.WebSocketHandler<{ origin: URL }> {
-    let debug = this.logDebug ? console.log : (msg: string) => {};
+    let debug = this.logDebug ? console.log : (_: string) => {};
     let wsRegisters = this.wsRegisters;
     let websocket: Bun.WebSocketHandler<{ origin: URL }> = {
       async message(ws, message: Buffer<ArrayBuffer>) {
